# server/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateLevelRequest:
    properties:
      description:
        example: Best of the best
        type: string
      name:
        example: Senior
        type: string
    required:
    - name
    type: object
  dto.CreateLocationRequest:
    properties:
      name:
        example: Chemnitz
        type: string
    required:
    - name
    type: object
  dto.CreateOrganizationRequest:
    properties:
      title:
        example: Staffbase
        type: string
    type: object
  dto.CreateOrganizationResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.DeleteOrganizationResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.GetOrganizationResponse:
    properties:
      id:
        type: string
      plan:
        type: string
      semantic:
        $ref: '#/definitions/model.OrganizationSemantic'
      status:
        type: string
      title:
        type: string
    type: object
...
  dto.LevelResponse:
    properties:
      message:
        type: string
    type: object
  dto.LocationResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.PatchOrganizationRequest:
    properties:
      title:
        example: StaffAlienBase
        type: string
    type: object
  dto.PatchOrganizationResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.UnauthorizedResponse:
    properties:
      message:
        type: string
    type: object
  dto.UpdateLevelRequest:
    properties:
      description:
        example: Living computer
        type: string
      name:
        example: Lvl 4
        type: string
    type: object
...
  lib.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact:
    email: msk.vitaly@gmail.com
    name: Vitalii Popov
    url: https://www.linkedin.com/in/mskVitalii/
  description: Semantic Contacts
  title: Semki
  version: "1.0"
paths:
...
  /api/v1/organization:
    delete:
      description: Deletes an organization from the MongoDB database by its ID.
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.DeleteOrganizationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletes an organization by its ID
      tags:
      - organizations
    get:
      description: Retrieves an organization from the MongoDB database by its ID.
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.GetOrganizationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieves an organization by user claims
      tags:
      - organizations
    patch:
      consumes:
      - application/json
      description: Updates only the provided fields of an organization
      parameters:
      - description: Fields to update
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.PatchOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.PatchOrganizationResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Partially updates an organization
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Creates a new organization in the MongoDB database.
      parameters:
      - description: Organization object to create
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.CreateOrganizationResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Creates a new organization
      tags:
      - organizations
  /api/v1/organization/levels:
    post:
      consumes:
      - application/json
      description: Creates a new level in the organization's semantic structure
      parameters:
      - description: Level data
        in: body
        name: level
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLevelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.LevelResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Adds a new level to organization
      tags:
      - organizations
  /api/v1/organization/levels/{levelId}:
    delete:
      description: Removes a level from organization by ID
      parameters:
      - description: Level ID
        in: path
        name: levelId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.LevelResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletes a level
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Updates level information by ID
      parameters:
      - description: Level ID
        in: path
        name: levelId
        required: true
        type: string
      - description: Level data to update
        in: body
        name: level
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateLevelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.LevelResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Updates a level
      tags:
      - organizations
  /api/v1/organization/locations:
    post:
      consumes:
      - application/json
      description: Creates a new location in the organization's semantic structure
      parameters:
      - description: Location data
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLocationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.LocationResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Adds a new location to organization
      tags:
      - organizations
  /api/v1/organization/locations/{locationId}:
    delete:
      description: Removes a location from organization by ID
      parameters:
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.LocationResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletes a location
      tags:
      - organizations
  /api/v1/organization/teams:
    post:
      consumes:
      - application/json
      description: Creates a new team in the organization's semantic structure
      parameters:
      - description: Team data
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.TeamResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Adds a new team to organization
      tags:
      - organizations
  /api/v1/organization/teams/{teamId}:
    delete:
      description: Removes a team from organization by ID
      parameters:
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.TeamResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deletes a team
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Updates team information by ID
      parameters:
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      - description: Team data to update
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/dto.TeamResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Updates a team
      tags:
      - organizations
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"